Extract to abstract class
